Default Refactor <broken>
- db connection get stuck after one use
