+++ .idea/sqldialects.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/webapps/WEB-INF/classes/DBservices/DatabaseOperations.java" dialect="GenericSQL" />
+    <file url="file://$PROJECT_DIR$/webapps/WEB-INF/classes/DBservices/DatabasesConnection.java" dialect="GenericSQL" />
     <file url="PROJECT" dialect="MySQL" />
   </component>
 </project>
