+++ webapps/META-INF/context.xml
@@ -6,7 +6,7 @@
             auth="Container"
             driverClassName="com.mysql.cj.jdbc.Driver"
 
-            maxActive="10"
+            maxActive="100"
             maxIdle="5"
             maxWait="10000"
